New scheduler
This scheduler is able to cancel tasks when the plugin is getting disabled.

+++ topographica-main/src/main/java/nl/rutgerkok/topographica/config/WorldConfig.java
@@ -51,7 +51,7 @@ public final class WorldConfig {
      *            Z position of the region.
      * @return True if it should be rendered, false otherwise.
      */
-    public boolean shouldRender(int regionX, int regionZ) {
+    public boolean shouldRender(long regionX, long regionZ) {
         if (radius == Integer.MAX_VALUE) {
             return true;
         }
